Styling and Bug fixes
UI: Fixed the visuals for the index, also did responsive work.
Fix: Passed content_type to form so we can use it as a title.
Fix: Made it possible for Title / Description / Slug to be empty to avoid errors because of the Indexer
Fix: added missing "location" to ParentIDTrait.
Fix: Link to related items was hardcoded with the taxonomy 'Keywords', made it dynamic.
Fix: unrelated, but added form_errors to content_type edit 
Tweak: Added Description to the indexer so we can use it in the table
Tweak: Added a pretty icon

Also, i found out that current category parent can only select top level categories, this should be able to go deeper.