Added MediaGalleryImageType
New: Added MediaGalleryImageType which can be set up for single or multiple select. Filetypes are defined in class.
New: Added Featured image to article, maybe we should move this to "Content/Content.php" to have it by default for all content
Tweak: Made mediagallery_selection dynamic so it works for relations and normal form fields