Final touches
You can use it on a content type page in the following way:

{% block javascripts %}
    {{ parent() }}
    {{ article.getMetaData()|get_social_media_scripts|raw }}
{% endblock %}

it checks which types of embeds there are, and only loads the required scripts. Since TinyMCE will not save script tags in the code, a different approach was needed. If there is another way to inject the right scripts into the body when needed. I'd love to know for a future update.